--- src/main/java/org/triumers/kmsback/Employee/command/Application/dto/CmdEmployeeDTO.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.command.Application.dto;
-
-public class CmdEmployeeDTO {
-}
