--- src/main/java/org/triumers/kmsback/Employee/command/Application/service/CmdEmployeeService.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.command.Application.service;
-
-public interface CmdEmployeeService {
-}
