--- src/main/java/org/triumers/kmsback/Employee/command/Application/service/CmdEmployeeServiceImpl.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.command.Application.service;
-
-public class CmdEmployeeServiceImpl implements CmdEmployeeService {
-}
