--- src/main/java/org/triumers/kmsback/Employee/command/domain/aggregate/entity/CmdEmployee.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.command.domain.aggregate.entity;
-
-public class CmdEmployee {
-}
