--- src/main/java/org/triumers/kmsback/Employee/command/domain/aggregate/vo/CmdResponseMessageVO.java
@@ -1,5 +1,0 @@
-package org.triumers.kmsback.Employee.command.domain.aggregate.vo;
-
-public class CmdResponseMessageVO {
-    private String message;
-}
