--- src/main/java/org/triumers/kmsback/Employee/command/domain/repository/CmdEmployeeRepository.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.command.domain.repository;
-
-public interface CmdEmployeeRepository {
-}
