--- src/main/java/org/triumers/kmsback/Employee/query/aggregate/entity/QryEmployee.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.query.aggregate.entity;
-
-public class QryEmployee {
-}
