--- src/main/java/org/triumers/kmsback/Employee/query/aggregate/vo/QryResponseMessageVO.java
@@ -1,5 +1,0 @@
-package org.triumers.kmsback.Employee.query.aggregate.vo;
-
-public class QryResponseMessageVO {
-    private String message;
-}
