--- src/main/java/org/triumers/kmsback/Employee/query/controller/QryEmployeeController.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.query.controller;
-
-public class QryEmployeeController {
-}
