--- src/main/java/org/triumers/kmsback/Employee/query/dto/QryEmployeeDTO.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.query.dto;
-
-public class QryEmployeeDTO {
-}
