--- src/main/java/org/triumers/kmsback/Employee/query/repository/QryEmployeeRepository.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.query.repository;
-
-public class QryEmployeeRepository {
-}
