--- src/main/java/org/triumers/kmsback/Employee/query/service/QryEmployeeService.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.query.service;
-
-public interface QryEmployeeService {
-}
