--- src/main/java/org/triumers/kmsback/Employee/query/service/QryEmployeeServiceImpl.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.Employee.query.service;
-
-public class QryEmployeeServiceImpl implements QryEmployeeService {
-}
