+++ src/main/java/org/triumers/kmsback/employee/command/Application/dto/CmdEmployeeDTO.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.command.Application.dto;
+
+public class CmdEmployeeDTO {
+}
