+++ src/main/java/org/triumers/kmsback/employee/command/Application/service/CmdEmployeeService.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.command.Application.service;
+
+public interface CmdEmployeeService {
+}
