+++ src/main/java/org/triumers/kmsback/employee/command/Application/service/CmdEmployeeServiceImpl.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.command.Application.service;
+
+public class CmdEmployeeServiceImpl implements CmdEmployeeService {
+}
