+++ src/main/java/org/triumers/kmsback/employee/command/domain/aggregate/entity/CmdEmployee.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.command.domain.aggregate.entity;
+
+public class CmdEmployee {
+}
