+++ src/main/java/org/triumers/kmsback/employee/command/domain/aggregate/vo/CmdResponseMessageVO.java
@@ -1,0 +1,5 @@
+package org.triumers.kmsback.employee.command.domain.aggregate.vo;
+
+public class CmdResponseMessageVO {
+    private String message;
+}
