+++ src/main/java/org/triumers/kmsback/employee/command/domain/repository/CmdEmployeeRepository.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.command.domain.repository;
+
+public interface CmdEmployeeRepository {
+}
