+++ src/main/java/org/triumers/kmsback/employee/query/aggregate/entity/QryEmployee.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.query.aggregate.entity;
+
+public class QryEmployee {
+}
