+++ src/main/java/org/triumers/kmsback/employee/query/controller/QryEmployeeController.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.query.controller;
+
+public class QryEmployeeController {
+}
