+++ src/main/java/org/triumers/kmsback/employee/query/dto/QryEmployeeDTO.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.query.dto;
+
+public class QryEmployeeDTO {
+}
