+++ src/main/java/org/triumers/kmsback/employee/query/repository/QryEmployeeRepository.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.query.repository;
+
+public class QryEmployeeRepository {
+}
