+++ src/main/java/org/triumers/kmsback/employee/query/service/QryEmployeeService.java
@@ -1,0 +1,4 @@
+package org.triumers.kmsback.employee.query.service;
+
+public interface QryEmployeeService {
+}
